--- src/main/java/gyeongdan/article/article/dto/ArticleUpdateRequest.java
@@ -1,15 +1,0 @@
-package gyeongdan.article.article.dto;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Data
-@Getter
-@Setter
-public class ArticleUpdateRequest {
-
-    private Long id;
-    private String title;
-    private String content;
-}
