--- src/main/java/gyeongdan/article/article/repository/ArticleJpaRepository.java
@@ -1,10 +1,0 @@
-package gyeongdan.article.article.repository;
-
-import gyeongdan.article.article.domain.Article;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ArticleJpaRepository extends JpaRepository<Article, Long> {
-
-}
