+++ src/main/java/gyeongdan/article/domain/ArticleRelatedDocuments.java
@@ -1,0 +1,31 @@
+package gyeongdan.article.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import gyeongdan.article.domain.Article;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+@Entity
+@Table(name = "article_related_documents", schema = "gyeongdan")
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@RequiredArgsConstructor
+public class ArticleRelatedDocuments {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @JsonIgnore
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "article_id")
+    private Article article;
+
+    private String title;
+    private String link;
+    private String snippet;
+}
