+++ src/main/java/gyeongdan/article/dto/ArticleUpdateRequest.java
@@ -1,0 +1,15 @@
+package gyeongdan.article.dto;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+@Data
+@Getter
+@Setter
+public class ArticleUpdateRequest {
+
+    private Long id;
+    private String title;
+    private String content;
+}
