+++ src/main/java/gyeongdan/article/dto/PopularArticleResponse.java
@@ -1,0 +1,14 @@
+package gyeongdan.article.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class PopularArticleResponse {
+    private Long id;
+    private String title;
+    private Long viewCount;
+}
