--- src/main/java/gyeongdan/article/recommend/repository/RecommendJpaRepository.java
@@ -1,9 +1,0 @@
-package gyeongdan.article.recommend.repository;
-
-import gyeongdan.article.recommend.domain.Recommends;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RecommendJpaRepository extends JpaRepository<Recommends, Long> {
-}
