--- src/main/java/gyeongdan/article/related_documents/domain/ArticleRelatedDocuments.java
@@ -1,31 +1,0 @@
-package gyeongdan.article.related_documents.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import gyeongdan.article.article.domain.Article;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-@Entity
-@Table(name = "article_related_documents", schema = "gyeongdan")
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class ArticleRelatedDocuments {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @JsonIgnore
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "article_id")
-    private Article article;
-
-    private String title;
-    private String link;
-    private String snippet;
-}
