+++ src/main/java/gyeongdan/article/repository/ArticleJpaRepository.java
@@ -1,0 +1,10 @@
+package gyeongdan.article.repository;
+
+import gyeongdan.article.domain.Article;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ArticleJpaRepository extends JpaRepository<Article, Long> {
+
+}
