+++ src/main/java/gyeongdan/article/service/ArticleRelatedDocumentsService.java
@@ -1,0 +1,21 @@
+package gyeongdan.article.service;
+
+import gyeongdan.article.domain.ArticleRelatedDocuments;
+import gyeongdan.article.repository.ArticleRelatedDocumentsRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ArticleRelatedDocumentsService {
+
+    private final ArticleRelatedDocumentsRepository articleRelatedDocumentsRepository;
+
+    public ArticleRelatedDocumentsService(ArticleRelatedDocumentsRepository articleRelatedDocumentsRepository) {
+        this.articleRelatedDocumentsRepository = articleRelatedDocumentsRepository;
+    }
+
+    public List<ArticleRelatedDocuments> getRelatedDocuments(Long articleId) {
+        return articleRelatedDocumentsRepository.findByArticleId(articleId);
+    }
+}
