+++ src/main/java/gyeongdan/shortform/dto/ShortFormDetailResponse.java
@@ -1,15 +1,10 @@
 package gyeongdan.shortform.dto;
 
-import gyeongdan.article.article.domain.Article;
-import gyeongdan.article.related_documents.domain.ArticleRelatedDocuments;
 import gyeongdan.shortform.domain.ShortForm;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 import java.sql.Timestamp;
-import java.util.List;
 
 @Setter
 @Getter
